requirements.txt
ccxt>=4.3.0
pandas>=2.2.2
# numpy 2.x pairs with TA-Lib 0.6.x
numpy>=2.0
pyyaml>=6.0.2
pydantic>=2.8.2
llama-cpp-python>=0.2.90
TA-Lib>=0.6.0
